Create minify-remove-htmlbodytags.php
This is a copy / paste of the original phpwee class "HtmlMin" with one important difference.  DOMDocument forces html, head, body and DOCtype upon you.  By inserting a wrapper at the top and then unwrapping everything outside of it at the bottom we work around that so we get back our original text minified with nothing added.